% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{The datasets identified as part of the review that formed the GPAD resource}
\format{
A data frame with 105 rows and 114 variables
}
\usage{
datasets
}
\description{
The dataset contains the datasets that were identified as part of the GPAD resource
}
\keyword{datasets}
