% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_map.R
\name{generate_world_map}
\alias{generate_world_map}
\title{generate_world_map}
\usage{
generate_world_map()
}
\value{
A dataset to be used to create a chloropleth world map
}
\description{
A function to generate the dataset used to create the chloropleth world map
}
