% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_country_prep.R
\name{multi_country_prep}
\alias{multi_country_prep}
\title{multi_country_prep}
\usage{
multi_country_prep(data = rv$datasets)
}
\arguments{
\item{data}{A dataframe - datasets}
}
\value{
A list with datasets that collect data in multiple countries split into individual countries
}
\description{
A function to split the datasets that have collected data in multiple countries into separate vectors
}
