% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_country_summary.R
\name{participant_country_summary}
\alias{participant_country_summary}
\title{participant_country_summary}
\usage{
participant_country_summary(data = rv$datasets)
}
\arguments{
\item{data}{A dataframe to be supplied to the function}
}
\value{
A dataframe summarising the number of participants collected in each country
}
\description{
A function to summarise the number of participants in each dataset
}
